Move appointment utils to check-in app root
