+++ src/applications/check-in/day-of/components/AppointmentDisplay/AppointmentAction.jsx
@@ -5,7 +5,7 @@ import { api } from '../../../api';
 
 import { URLS } from '../../utils/navigation';
 import { useFormRouting } from '../../../hooks/useFormRouting';
-import { ELIGIBILITY, areEqual } from '../../utils/appointment/eligibility';
+import { ELIGIBILITY, areEqual } from '../../../utils/appointment/eligibility';
 import recordEvent from 'platform/monitoring/record-event';
 import format from 'date-fns/format';
