+++ src/applications/check-in/utils/appointment/eligibility.js
