+++ src/applications/check-in/utils/appointment/index.js
