+++ src/applications/check-in/utils/appointment/status.test.unit.spec.js
